Enigma OO Update
Made the Enigma Machine itself an object instead of having a ton of static variables.

I also implemented a method for the rotors to rotate themselves by a massive number without having to iteratively rotate by 1. Get that O(n) out of here, we ain't got no place for that.